% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{diff_in_means}
\alias{diff_in_means}
\title{Calculate difference in means}
\usage{
diff_in_means(df, group_col, outcome_col, treatment_value = NULL)
}
\arguments{
\item{df}{A data frame}

\item{group_col}{The name of the column in df that corresponds to the group label}

\item{outcome_col}{The name of the column in df that corresponds to the outcome variable}

\item{treatment_value}{The value of group_col to be considered 'treatment'}
}
\value{
The difference in mean outcome between the two groups
}
\description{
This function takes a data frame, and group and outcome column names as input
and returns the difference in mean outcome between the two groups
}
