% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{npc}
\alias{npc}
\title{Run NPC}
\usage{
npc(
  df,
  group_col,
  outcome_cols,
  strata_col = NULL,
  test_stat = "diff_in_means",
  perm_func = permute_group,
  combn = "fisher",
  shift = 0,
  reps = 10000,
  perm_set = NULL,
  complete_enum = F,
  seed = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{group_col}{The name of the column in df that corresponds to the group label}

\item{outcome_cols}{The names of the columns in df that corresponds to the outcome variable}

\item{strata_col}{The name of the column in df that corresponds to the strata}

\item{test_stat}{Test statistic function}

\item{perm_func}{Function to permute group}

\item{combn}{Combining function method to use}

\item{shift}{Value of shift to apply in one- or two-sample problem}

\item{reps}{Number of iterations to use when calculating permutation p-value}

\item{perm_set}{Matrix of permutations to use instead of reps iterations of perm_func}

\item{complete_enum}{Boolean, whether to calculate P-value under complete enumeration of permutations}

\item{seed}{An integer seed value}
}
\value{
A list containing the permutation test p-value, and the test statistic distribution if applicable
}
\description{
This function takes a data frame and group and outcome column names as input
and returns the NPC omnibus p-value
}
