% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{permutation_test}
\alias{permutation_test}
\title{Run permutation test}
\usage{
permutation_test(
  df,
  group_col,
  outcome_col,
  strata_col = NULL,
  test_stat = "diff_in_means",
  perm_func = permute_group,
  alternative = "two-sided",
  shift = 0,
  reps = 10000,
  perm_set = NULL,
  return_test_dist = FALSE,
  return_perm_dist = FALSE,
  seed = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{group_col}{The name of the column in df that corresponds to the group label}

\item{outcome_col}{The name of the column in df that corresponds to the outcome variable}

\item{strata_col}{The name of the column in df that corresponds to the strata}

\item{test_stat}{Test statistic function}

\item{perm_func}{Function to permute group}

\item{alternative}{String, two-sided or one-sided (greater or less) p-value}

\item{reps}{Number of iterations to use when calculating permutation p-value}

\item{perm_set}{Matrix of permutations to use instead of reps iterations of perm_func}

\item{seed}{An integer seed value}

\item{return_dist}{Boolean, whether to return test statistic distribution under permutations}
}
\value{
A list containing the permutation test p-value, and the test statistic distribution if applicable
}
\description{
This function takes a data frame and group and outcome column name as input
and returns the dataframe with the group column randomly permuted
}
